style: mv x509 auth around
